src/main/java/net/lemixeur/domain/user/UniqueEmailValidator.java
package net.lemixeur.domain.user;

import javax.validation.*;

import net.lemixeur.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

@Component
public class UniqueEmailValidator implements ConstraintValidator<UniqueEmail, String> {

    @Autowired
    private UserService userService;

    @Override
    public boolean isValid(String value, ConstraintValidatorContext context) {
        return !userService.isEmailAlreadyInUse(value);
    }

}